final version? slow in random cities. must change to original but is just uncomment stuff. when this is going to end tho ?

+++ node.hpp
@@ -22,10 +22,10 @@ struct Node_h {
 };
 
 struct Solution {
-    int g1;
+    double g1;
     double g2;
     
-    Solution(int i, double j): g1(i), g2(j){}
+    Solution(double i, double j): g1(i), g2(j){}
     ~Solution(){}
 };
